added codegen config
