+++ python2verilog/api/context.py
@@ -5,9 +5,11 @@ Functions that take text as input
 
 import copy
 import logging
+from typing import Optional
 
 from python2verilog import ir
 from python2verilog.backend import verilog
+from python2verilog.backend.verilog.config import CodegenConfig
 from python2verilog.frontend.generator2ir import Generator2Graph
 from python2verilog.optimizer.optimizer import OptimizeGraph
 from python2verilog.utils.assertions import get_typed
@@ -29,7 +31,7 @@ def context_to_codegen(context: ir.Context):
     return verilog.CodeGen(ir_root, context), ir_root
 
 
-def context_to_verilog(context: ir.Context) -> tuple[str, str]:
+def context_to_verilog(context: ir.Context, config: CodegenConfig) -> tuple[str, str]:
     """
     Converts a context to a verilog module and testbench
 
@@ -39,7 +41,7 @@ def context_to_verilog(context: ir.Context) -> tuple[str, str]:
     ver_code_gen, _ = context_to_codegen(context)
 
     module_str = ver_code_gen.get_module_str()
-    tb_str = ver_code_gen.get_testbench_str()
+    tb_str = ver_code_gen.get_testbench_str(random_ready=config.random_ready)
 
     return module_str, tb_str
