+++ python2verilog/backend/verilog/config.py
@@ -1,0 +1,15 @@
+"""
+Configurations
+"""
+
+from dataclasses import dataclass
+
+
+@dataclass
+class CodegenConfig:
+    """
+    Configurations
+    """
+
+    # Enable random ready signal in testbench
+    random_ready: bool = False
